Use hardhat instead of truffle in graphical test script.
Also improves a few things like using `pgrep`.

+++ scripts/send-back-and-forth-test.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 # Test sending doge to eth and back
 # Tested on Mac OSX only
 
@@ -11,8 +13,10 @@ agentCodeDir=/path/agentCodeDir
 agentDataDir=/path/agentDataDir
 dogethereumContractsCodeDir=/path/dogethereumContractsCodeDir
 
+NETWORK="integrationDogeRegtest"
+
 # Print instructions on the console
-set -o xtrace
+set -o xtrace -o nounset
 # Stop dogecoin-qt
 killall $dogecoinQtProcessName
 sleep 3s
@@ -25,31 +29,31 @@ sleep 10s
 # Mine a doge block so dogecoin-qt is fully up and running
 curl --user $dogecoinQtRpcuser:$dogecoinQtRpcpassword  --data-binary '{"jsonrpc": "1.0", "id":"curltest", "method": "generate", "params": [1] }' -H 'content-type: text/plain;' http://127.0.0.1:41200/
 # Clear agent data dir
-rm -rf $agentDataDir/*
+# TODO: is this necessary?
+rm -rf ${agentDataDir:?}/*
 # Restart ganache
-kill $(ps aux | grep ganache | grep -v grep | awk '{print $2}')
+kill "$(pgrep ganache)"
 sleep 1s
-ganache-cli --gasLimit 400000000000 > ganachelog.txt &
-# Remove compiled contract to force recompiling
-rm -rf $dogethereumContractsCodeDir/build/contracts/
+npm run ganache > ganachelog.txt &
 # Compile and deploy contracts
-truffle deploy --network integrationDogeRegtest | grep Error
+npx hardhat compile --quiet
+npx hardhat run --network $NETWORK scripts/deployDogethereum.ts
 # Init contracts: initial doge header and operator
-truffle exec  --network integrationDogeRegtest scripts/init_contracts_local.js 
+npx hardhat run --network $NETWORK scripts/init_contracts_local.ts
 # Print debug.js status
-truffle exec  --network integrationDogeRegtest scripts/debug.js 
+npx hardhat run --network $NETWORK scripts/debug.ts
 echo "Please, start the agent..."
 # Wait for agent to relay doge lock tx to eth and dogetokens minted
-truffle exec  --network integrationDogeRegtest scripts/wait_token_balance.js 
+npx hardhat run --network $NETWORK scripts/wait_token_balance.ts
 # Prepare sender address to do unlocks
-truffle exec  --network integrationDogeRegtest scripts/prepare_sender.js
-for i in $(seq 1 2); do 
-	# Send eth unlock tx 
+npx hardhat run --network $NETWORK scripts/prepare_sender.ts
+for i in {1..2}; do
+	# Send eth unlock tx
 	node ../dogethereum-tools/user/unlock.js --ethnetwork ganacheDogeRegtest --json $dogethereumContractsCodeDir/build/contracts/DogeToken.json --sender 0xd2394f3fad76167e7583a876c292c86ed10305da --receiver ncbC7ZY1K9EcMVjvwbgSBWKQ4bwDWS4d5P --value 300000000
 	# Print debug.js status
-	truffle exec  --network integrationDogeRegtest scripts/debug.js 
-	# Mine 5 eth blocks so unlock eth tx has enought confirmations
-	for j in $(seq 1 5); do 
+	npx hardhat run --network $NETWORK scripts/debug.ts
+	# Mine 5 eth blocks so unlock eth tx has enough confirmations
+	for j in {1..5}; do
 		curl -X POST --data '{"jsonrpc":"2.0","method":"evm_mine","params":[],"id":74}' http://localhost:8545;
 	done
 	# Wait for Eth to Doge agent to sign and broadcast doge unlock tx
@@ -57,5 +61,5 @@ for i in $(seq 1 2); do
 	# Mine 10 doge blocks so doge unlock tx has enough confirmations
 	curl --user $dogecoinQtRpcuser:$dogecoinQtRpcpassword  --data-binary '{"jsonrpc": "1.0", "id":"curltest", "method": "generate", "params": [10] }' -H 'content-type: text/plain;' http://127.0.0.1:41200/
 	# Wait for agent to relay doge unlock tx to eth and utxo length updated
-	truffle exec  --network integrationDogeRegtest scripts/wait_two_utxos.js 
+	npx hardhat run --network $NETWORK scripts/wait_two_utxos.ts
 done
